refactor: 스텝1 수정사항 반영
