+++ OpenMarket/OpenMarket/Data/OpenMarketLoadData.swift
@@ -8,15 +8,12 @@
 import Foundation
 
 struct OpenMarketLoadData {
-    
-    let networkManager: NetworkManager
-    
     func requestOpenMarketMainPageData(page: String, completionHandler: @escaping (OpenMarketItems) -> ()) {
         guard let url = URL(string: String(describing: OpenMarketUrl.listLookUp) + page) else {
             return
         }
         
-        networkManager.request(httpMethod: .get, url: url, body: nil, .json) { result in
+        NetworkManager.request(httpMethod: .get, url: url, body: nil, .json) { result in
             switch result {
             case .success(let data):
                 guard let jsonItem = try? ParsingManager.jsonDecode(data: data, type: OpenMarketItems.self) else { return }
