+++ OpenMarket/OpenMarket/Model/NetworkManager.swift
@@ -16,66 +16,21 @@ enum NetworkError: Error {
 
 //MARK:-NetworkManager
 struct NetworkManager {
-    
-    //MARK: HTTPBody Parameter Type
-    typealias requestBodyType = [String: Any]
-    
+ 
     //MARK: Properties
-    let session: URLSession
-    private let boundary = "Boundary\(UUID().uuidString)"
+    static var session: URLSession = URLSession.shared
     
     //MARK: Private Method
-    private func makeContentDispositionLine(contentType: ContentType) -> String {
-        return "Content-Disposition: \(contentType); "
-    }
-    
-    private func createRequest(httpMethod: HTTPMethod, url: URL, body: Data?, _ contentType: ContentType) -> URLRequest {
+    private static func createRequest(httpMethod: HTTPMethod, url: URL, body: Data?, _ contentType: ContentType) -> URLRequest {
         var request = URLRequest(url: url)
-        request.setValue("\(contentType); boundary=\(boundary)", forHTTPHeaderField: "Content-Type")
+        request.setValue("\(contentType); boundary=\(Boundary.literal)", forHTTPHeaderField: "Content-Type")
         request.httpMethod = String(describing: httpMethod)
         request.httpBody = body
         return request
     }
     
     //MARK: Method
-    func createHTTPBody(contentType: ContentType, with parameters: requestBodyType?, media: [Media]?) -> Data? {
-        let lineBreak = "\r\n"
-        var body = Data()
-        
-        if let parameters = parameters {
-            for (key, value) in parameters {
-                body.append(boundary + lineBreak)
-                
-                switch contentType {
-                case .json:
-                    body.append("\(makeContentDispositionLine(contentType: .json))\(lineBreak)")
-                    guard let jsonData = value as? Data else {
-                        return nil
-                    }
-                    body.append(jsonData)
-                case .multipart:
-                    body.append("\(makeContentDispositionLine(contentType: .multipart))name=\"\(key)\"\(lineBreak)")
-                    body.append("\(value)\(lineBreak)")
-                }
-            }
-        }
-        
-        if let media = media {
-            for image in media {
-                body.append(boundary + lineBreak)
-                body.append("\(makeContentDispositionLine(contentType: .multipart))name=\"\(image.key)\"; filename=\"\(image.fileName)\"\(lineBreak)")
-                body.append("Content-Type: \(image.mimeType)")
-                body.append(image.data)
-                body.append(lineBreak)
-            }
-        }
-        
-        body.append("--\(boundary)--\(lineBreak)")
-        
-        return body
-    }
-    
-    func request(httpMethod: HTTPMethod, url: URL, body: Data?, _ contentType: ContentType, _ completionHandler: @escaping (Result<Data, NetworkError>) -> ()) {
+        static func request(httpMethod: HTTPMethod, url: URL, body: Data?, _ contentType: ContentType, _ completionHandler: @escaping (Result<Data, NetworkError>) -> ()) {
         let request = createRequest(httpMethod: httpMethod, url: url, body: body, contentType)
         
         session.dataTask(with: request) { (data, response, error) in
